requirements.txt
mcp[cli]
asyncpg
python-dotenv

